+++ src/harlequin/actions.py
@@ -247,6 +247,9 @@ HARLEQUIN_ACTIONS = {
     "results_viewer.copy_selection": Action(
         target=ResultsTable, action="copy_selection"
     ),
+    "results_viewer.view_cell": Action(
+        target=ResultsTable, action="view_cell", description="View Cell"
+    ),
     "results_viewer.select_cursor": Action(target=ResultsTable, action="select_cursor"),
     # Moving the cursor
     "results_viewer.cursor_up": Action(target=ResultsTable, action="cursor_up"),